cleaned up and split api files by routes
