+++ api/routes/projects.api.js
@@ -1,0 +1,43 @@
+//jshint esversion:6
+const express = require('express');
+const router = express.Router();
+const Projects = require('../models/projects.model');
+
+router.post('/newProject', function (req, res, next) {
+    console.log('saving new project');
+
+    console.log(req.body);
+
+    // received remotePermitted as string - convert it to boolean
+    var remotePermittedConverted = (req.body.remotePermitted === "true") ? true : false;
+
+    // build the constructor
+    var newProject = new Projects({
+        projectName: req.body.projectName,
+        projectDescription: req.body.projectDescription,
+        projectStartDate: req.body.projectStartDate,
+        projectEndDate: req.body.projectEndDate,
+        projectLocation: req.body.projectLocation,
+        projectCategoryByCollege: req.body.projectCategoryByCollege,
+        projectCategoryByProgram: req.body.projectCategoryByProgram,
+        otherSkillsDesired: req.body.otherSkillsDesired,
+        remotePermitted: remotePermittedConverted,
+        _primaryProjectOwner: req.body._primaryProjectOwner,
+        _usersInvited: null,
+        _usersAssigned: null
+    });
+
+    newProject.save(function (err, newProject) {
+        if (err) {
+            console.log("error saving project");
+            console.log(err);
+            res.send("posting error");
+        } else {
+            console.log('posted project');
+            res.send('posted project');
+        }
+    });
+
+});
+
+module.exports = router;
